Move helmclient and logger under pkg
